+++ system/info_darwin.go
@@ -11,6 +11,7 @@ package system
 
 import (
 	"errors"
+	"os/exec"
 	"strings"
 	"syscall"
 )
@@ -22,37 +23,55 @@ func GetSystemInfo() (*SystemInfo, error) {
 	hostname, err := syscall.Sysctl("kern.hostname")
 
 	if err != nil || hostname == "" {
-		return &SystemInfo{}, errors.New("Can't read hostname info")
+		return nil, errors.New("Can't read hostname info")
 	}
 
 	os, err := syscall.Sysctl("kern.ostype")
 
 	if err != nil || os == "" {
-		return &SystemInfo{}, errors.New("Can't read os info")
+		return nil, errors.New("Can't read os info")
 	}
 
 	kernel, err := syscall.Sysctl("kern.osrelease")
 
 	if err != nil || kernel == "" {
-		return &SystemInfo{}, errors.New("Can't read kernel info")
+		return nil, errors.New("Can't read kernel info")
 	}
 
 	arch, err := syscall.Sysctl("kern.version")
 
 	if err != nil || arch == "" {
-		return &SystemInfo{}, errors.New("Can't read arch info")
+		return nil, errors.New("Can't read arch info")
 	}
 
 	archSlice := strings.Split(arch, "/")
 
 	if len(archSlice) != 2 {
-		return &SystemInfo{}, errors.New("Can't read arch info")
+		return nil, errors.New("Can't read arch info")
 	}
 
+	cleanArch := strings.ToLower(strings.Replace(archSlice[len(archSlice)-1], "RELEASE_", "", -1))
+
 	return &SystemInfo{
-		hostname,
-		os,
-		kernel,
-		strings.ToLower(strings.Replace(archSlice[len(archSlice)-1], "RELEASE_", "", -1)),
+		Hostname:     hostname,
+		OS:           os,
+		Distribution: DARWIN_OSX,
+		Version:      getOSXVersion(),
+		Kernel:       kernel,
+		Arch:         cleanArch,
 	}, nil
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+func getOSXVersion() string {
+	cmd := exec.Command("sw_vers", "-productVersion")
+
+	versionData, err := cmd.Output()
+
+	if err != nil {
+		return ""
+	}
+
+	return string(versionData)
 }